Board now tracks connections
Adjusted boardclass.py so that it checks for and tracks connections.

This reduces the amount of iteration during win checks (for l in connections: if len(l) > 5)
and is also very helpful for drawing the winning streak to the board.

These optimizations will save a butt load when minmax gets implemented.