Refactored connection checking
The previous system had a couple of loopholes, especially when it came to dead_connections. Now the code is much more readable and functions more robustly. There are some exploits with no precondition checking (if you check cords out of bounds, etc), but those do not occur with user input, only if testing from the console.

Adjust main.py to interface with the new methods.