MinMax Implemented
static_eval and recursive min & max search implemented

there is currently an issue where invalid moves can be returned

Board class now has a handle_turn method that changes team, makes a move, etc, so as to simplify turn-making. It also returns depending on the succss of a move. Also found a bug in the connection system that was patched.

Added a tester to test the min-max