animation.py created
All animation / pygame functionality has been taken out of main.py and put into animation.py . Later (hopefully) there will also be SFX handling. I do not really like how it is, but what is now is definitely a step in the right direction. The big slop is that it's unknown what calls are made when, and there's the recchid smell of passing pygame imported data across functions between files. Otherwise, I think this is good.