Min-Max finished
The min-max search now terminates when a win is found, speeding up end-game calculations and (more importantly) making the AI always win as fast as possible. Also adjusted static-evals so that all wins are treated equally. This means that a board with a 4-long streak for both sides will be declared as a winner for one side, but given that those are not possible given the game-loop, this is ok.

Min-max also now checks whether or not a move is valid, speeding up late-game

Min-max now looks at moves from the center outwards, in the following order (3,4,2,5,1,6,0)

main.py no longer prints "calculating move" and "found: {0}"