GUI now functional, prep for animation done
At the end of the game, a gui now pops up with two options - reset and quit. Previous selected is now used for when the button is pressed. No longer draws preview ghosts if a win.

New button state (and image) for deselected. Images have been renamed.

set_piece_color(board) will now align the current color to be that of the board's, assuming red = player turn

draw_square_of_piece(..) now can be called to redraw a section of the board (for animation, not currently in use)