EVERYTHING DONE :D
The GUI now has its own falling animation, and is completely correctly functional. main.py now recognizes ties, and a couple functions have been tweaked in display.py

Victory sound added for winning, and game over sound added to signal the game is over - it's not really a sad sound. README now has a legal section to mention that all sounds are CC-0

animations should be put into a seperate animation.py but it's ok because things are over

Yes, the TODO is not completely checked off but I'm calling it a day. No two player mode, or victory sound effect, or anything of the sort - but at least it's a functioning ai. At the end of the day, this was pretty good. There is a non existent animation for resetting the game. I feel like I could work on this for a long long time